Cleanup some dataset inheritance and merge issues
