+++ polaris/experimental/_dataset_v2.py
@@ -14,7 +14,7 @@ from polaris.dataset._adapters import Adapter
 from polaris.dataset._base import BaseDataset
 from polaris.dataset.zarr._manifest import calculate_file_md5, generate_zarr_manifest
 from polaris.utils.errors import InvalidDatasetError
-from polaris.utils.types import AccessType, HubOwner, ZarrConflictResolution
+from polaris.utils.types import AccessType, ChecksumStrategy, HubOwner, ZarrConflictResolution
 
 _INDEX_ARRAY_KEY = "__index__"
 
@@ -45,7 +45,6 @@ class DatasetV2(BaseDataset):
     version: ClassVar[Literal[2]] = 2
     _zarr_manifest_path: str | None = PrivateAttr(None)
     _zarr_manifest_md5sum: str | None = PrivateAttr(None)
-    _md5sum: str | None = PrivateAttr(None)
 
     # Redefine this to make it a required field
     zarr_root_path: str
@@ -94,7 +93,7 @@ class DatasetV2(BaseDataset):
         """Return all row indices for the dataset
 
         Warning: Memory consumption
-            This feature is added for completeness' sake, but when datasets get large could consume a lot of memory.
+            This feature is added for completeness' sake, but it should be noted that large datasets could consume a lot of memory.
             E.g. storing a billion indices with np.in64 would consume 8GB of memory. Use with caution.
         """
         return np.arange(len(self), dtype=int)
@@ -146,31 +145,6 @@ class DatasetV2(BaseDataset):
     def has_zarr_manifest_md5sum(self) -> bool:
         """Whether the md5sum for this dataset's zarr manifest has been computed and stored."""
         return self._zarr_manifest_md5sum is not None
-
-    @computed_field
-    @property
-    def md5sum(self) -> str:
-        """
-        Lazily compute the checksum once needed.
-
-        The checksum of the DatasetV2 is computed from the Zarr Manifest and is _not_ deterministic.
-        """
-        if not self.has_md5sum:
-            logger.info("Computing the checksum. This can be slow for large datasets.")
-            self.md5sum = calculate_file_md5(self.zarr_manifest_path)
-        return self._md5sum
-
-    @md5sum.setter
-    def md5sum(self, value: str):
-        """Set the checksum."""
-        if not re.fullmatch(r"^[a-f0-9]{32}$", value):
-            raise ValueError("The checksum should be the 32-character hexdigest of a 128 bit MD5 hash.")
-        self._md5sum = value
-
-    @property
-    def has_md5sum(self) -> bool:
-        """Whether the md5sum for this class has been computed and stored."""
-        return self._md5sum is not None
 
     def get_data(self, row: int, col: str, adapters: dict[str, Adapter] | None = None) -> np.ndarray | Any:
         """Indexes the Zarr archive.
@@ -278,21 +252,22 @@ class DatasetV2(BaseDataset):
             json.dump(serialized, f)
         return dataset_path
 
-    def cache(self, verify_checksum: bool = True) -> str:
+    def cache(self) -> str:
         """Caches the dataset by downloading all additional data for pointer columns to a local directory.
-
-        Args:
-            verify_checksum: Whether to verify the checksum of the dataset after caching.
 
         Returns:
             The path to the cache directory.
         """
-        # NOTE (cwognum): We don't support a deterministic checksum for the Dataset V2 yet,
-        #  so verification doesn't make sense. See also:
-        #  https://github.com/polaris-hub/polaris/issues/188
-        return super().cache(verify_checksum=False)
+        self.to_json(self._cache_dir, load_zarr_from_new_location=True)
+        return self._cache_dir
 
     def _repr_dict_(self) -> dict:
         """Utility function for pretty-printing to the command line and jupyter notebooks"""
         repr_dict = self.model_dump()
         return repr_dict
+
+    def should_verify_checksum(self, strategy: ChecksumStrategy) -> bool:
+        """
+        Determines whether to verify the checksum of the dataset based on the strategy.
+        """
+        return False
